requirements.txt
chainlit>=0.7.700
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
langchain-community>=0.0.13